mass: fix broken typed component API + add ECS regression suite
Rotating the loop to test coverage: cardinal::mass (the data-oriented
ECS behind crowds/swarms/projectiles) had ZERO tests. Added a
deterministic characterization suite (tests/mass, 67 checks, wired
into the sweep — TOTAL 56 -> 57): component registration, entity
lifecycle + free-list reuse, add/remove/has/get with archetype moves
preserving shared components, for_each mask filtering, for_each_chunk
typed slices, and the 1024-entity chunk split + Stats.

The new coverage immediately caught a serious latent bug: the typed
API was entirely non-functional. register_component<T> keyed
type_to_bit with static_cast<u32>(typeid(T).hash_code()) (truncated),
while add<T>/get<T> look the type up via bit_for_type_(typeid(T).
hash_code()) un-truncated. On any platform whose hash_code() has
high bits set (MSVC), the keys never matched, so every typed add<T>
returned false and every get<T> returned nullptr — the ergonomic
Mass API silently did nothing. type_hash is u64 end-to-end; the fix
drops the truncating cast so register and lookup use the same key.

(Bit-index registration via the raw register_component path was fine,
which is why nothing caught it until typed coverage existed.) Test
get<>() derefs are pointer-guarded so a future regression FAILs
cleanly instead of crashing.

Release build clean /W4 /WX, 57/57 regression green.

+++ runtime/mass/include/cardinal/mass/mass.hpp
@@ -55,9 +55,15 @@ public:
     // ----- Component registration -----------------------------------
     template <class T>
     u32 register_component(const cardinal::string& name) {
+        // NOTE: pass the FULL 64-bit hash. add<T>/get<T> look it up via
+        // bit_for_type_(typeid(T).hash_code()) un-truncated; an earlier
+        // static_cast<u32> here keyed type_to_bit with a truncated hash,
+        // so on any platform whose hash_code() has high bits set (MSVC)
+        // every typed add<T>/get<T> silently missed and returned
+        // false/nullptr. type_hash is u64 end-to-end.
         return register_component_internal_(
             name, static_cast<u32>(sizeof(T)), static_cast<u32>(alignof(T)),
-            static_cast<u32>(typeid(T).hash_code()));
+            typeid(T).hash_code());
     }
     u32 component_count() const noexcept;
     const ComponentDesc* describe_component(u32 bit) const noexcept;
